feat(bulk-upload): validate zipcode using regex
Client side bulk upload was not validating zipcode. This was causing
an unchecked zipcode to be denied on the server instead of the client,
causing a label on weship to be uploaded, but not on our website. I
have implemented the validation so that they will get denied before it
makes it to the server.