feat(duoplane): add state to track partials ships
At the stage its at now, you can click on the dropdown which works with
no animation. It then pops up an add shipment button. Once the add
shipment button is clicked, it will add a partial shipment to that
specific PO. Still needs styling, the inputs aren't bound to state, and
there are several bugs