feat(duoplane): finish state addition logic
At this point, the form can fully add state. When creating shipments, it
checks to see if the po already exists in the shipments. If it does, it
just adds another partial shipment to the existing po. If it doesn't, it
will initialize that po in the shipment along with a single partial
shipment