feat(duoplane-fetch): add error handling to fetch
Added custom errors for different error situations. An error can be
thrown in a couple different ways. If a status code is returned but it
was 400-500 range, we give errors that will show up on the client without
redirecting the user. If a status code isn't returned, we redirect the
user to the error page with the custom error stating what went wrong.
Every other error just takes them to error page.