feat(duoplane): add trpc route for get duoplane
TRPC route is meant to fetch data from duoplane. This is so we can
dynamically fetch from the front end (passing in pages to our fetch
request) while being able to hide the key on the backend. Right now this
is super slow, so I need to figure out a way to optimize the backend
call. In the future, we will store the duoplane key in the backend and
get it before we fetch from duoplane.