fix(label-creation): fix bulk label pricing bug
There was a bug in the createLabels function that was setting the price
it costs us (proglo shipping) as the prices that get stored in our db
for how much labels cost our users. This is incorrect, and we should be
storing the price it cost the user to create a label in the db. A side
effect of this bug caused users who use bulk uploader to have the same
cost as us when buying labels. This means that all labels that were bulk
uploaded, we never got any profit on them. Once I updated that, I
created better error handling in all the places that function is being
called.