feat(duoplane): create submitPoOrders function
Function is attached to submit button. It will create the labels with
duoplane, upload that info to our db, and then upload it to duoplane.
Lot of error handling was added, but nothing has been done with the
errors yet. I want to display errors to the user so they know where in
the process something went wrong if anything went wrong.