fix(single-label): update weight calculation func
The updateWeight function was calculating the user price and throwing an
uncaught error every time the input was blank or not between 0-70. I
updated the function to use the existing calculateCost function, and
catch any errors and render them to the UI. This way, the user knows if
they need to change their weight, but also prevents unnecessary errors
going to sentry.